requirements.txt
flask
flask-cors
requests
python-dotenv
gunicorn
openai
